pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mp3-split"
version = "0.1.0"
description = ""
authors = ["Daniel Teichman"]
license = "GPLv3"
readme = "README.md"
packages = [{include = "mp3_split"}]

[tool.poetry.dependencies]
python = "^3.7.2"
pyyaml = "^6.0"
ffmpeg-python = "^0.2.0"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
pylint = "^2.16.2"

[tool.poetry.scripts]
mp3-split = "mp3_split:main"
